old-back-end/requirements.txt
scikit-learn
numpy
xgboost
pandas
gunicorn
Flask
Flask-SQLAlchemy
Flask-Babel
Flask-JSON
flask-JWT-Extended
Werkzeug
PyMySQL
